third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/mixedContent/websocket/insecure-websocket-in-secure-page-worker-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-throw-on-construction-worker.html' was loaded over HTTPS, but attempted to connect to the insecure WebSocket endpoint 'ws://example.test:8880/echo'. This request has been blocked; this endpoint must be available over WSS.
This is a testharness.js-based test.
PASS Test that an insecure WebSocket connection is blocked when created in a worker created from HTTPS resource. 
Harness: the test ran to completion.
